完善JWT
