+++ febs-auth/src/main/java/cc/mrbird/febs/auth/controller/SecurityController.java
@@ -2,8 +2,6 @@ package cc.mrbird.febs.auth.controller;
 
 import cc.mrbird.febs.auth.manager.UserManager;
 import cc.mrbird.febs.auth.service.ValidateCodeService;
-import cc.mrbird.febs.common.entity.FebsResponse;
-import cc.mrbird.febs.common.entity.system.SystemUser;
 import cc.mrbird.febs.common.exception.ValidateCodeException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -30,16 +28,8 @@ public class SecurityController {
         return principal;
     }
 
-    @GetMapping("user/detail")
-    public FebsResponse currentUserDetail(Principal principal) {
-        SystemUser user = userManager.findByName(principal.getName());
-        user.setPassword("secret");
-        return new FebsResponse().data(user);
-    }
-
     @GetMapping("captcha")
     public void captcha(HttpServletRequest request, HttpServletResponse response) throws IOException, ValidateCodeException {
         validateCodeService.create(request, response);
     }
-
 }
